+++ ompi/mca/common/ompio/common_ompio_file_open.c
@@ -110,7 +110,7 @@ int mca_common_ompio_file_open (ompi_communicator_t *comm,
 
     ompio_fh->f_get_num_aggregators=mca_io_ompio_get_num_aggregators;
     ompio_fh->f_get_bytes_per_agg=mca_io_ompio_get_bytes_per_agg;
-    ompio_fh->f_set_aggregator_props=ompi_io_ompio_set_aggregator_props;
+    ompio_fh->f_set_aggregator_props=mca_io_ompio_set_aggregator_props;
 
     /* This fix is needed for data seiving to work with
        two-phase collective I/O */
