+++ ompi/mca/io/ompio/io_ompio_file_open.c
@@ -516,94 +516,3 @@ int mca_io_ompio_file_get_position_shared (ompi_file_t *fp,
     return ret;
 }
 
-int mca_io_ompio_cart_based_grouping(mca_io_ompio_file_t *ompio_fh)
-{
-    int k = 0;
-    int j = 0;
-    int n = 0;
-    int tmp_rank = 0;
-    int coords_tmp[2] = { 0 };
-
-    cart_topo_components cart_topo;
-
-    ompio_fh->f_comm->c_topo->topo.cart.cartdim_get(ompio_fh->f_comm, &cart_topo.ndims);
-
-    cart_topo.dims = (int*)malloc (cart_topo.ndims * sizeof(int));
-    if (NULL == cart_topo.dims) {
-        opal_output (1, "OUT OF MEMORY\n");
-        return OMPI_ERR_OUT_OF_RESOURCE;
-    }
-    cart_topo.periods = (int*)malloc (cart_topo.ndims * sizeof(int));
-    if (NULL == cart_topo.periods) {
-        opal_output (1, "OUT OF MEMORY\n");
-        return OMPI_ERR_OUT_OF_RESOURCE;
-    }
-    cart_topo.coords = (int*)malloc (cart_topo.ndims * sizeof(int));
-    if (NULL == cart_topo.coords) {
-        opal_output (1, "OUT OF MEMORY\n");
-        return OMPI_ERR_OUT_OF_RESOURCE;
-    }
-
-    ompio_fh->f_comm->c_topo->topo.cart.cart_get(ompio_fh->f_comm,
-                                                 cart_topo.ndims,
-	                                         cart_topo.dims,
-	                                         cart_topo.periods,
-  	                                         cart_topo.coords);
-
-    ompio_fh->f_init_procs_per_group = cart_topo.dims[1]; //number of elements per row
-    ompio_fh->f_init_num_aggrs = cart_topo.dims[0];  //number of rows
-
-    //Make an initial list of potential aggregators
-    ompio_fh->f_init_aggr_list = (int *) malloc (ompio_fh->f_init_num_aggrs * sizeof(int));
-    if (NULL == ompio_fh->f_init_aggr_list) {
-        opal_output (1, "OUT OF MEMORY\n");
-        return OMPI_ERR_OUT_OF_RESOURCE;
-    }
-
-    for(k = 0; k < cart_topo.dims[0]; k++){
-        coords_tmp[0] = k;
-        coords_tmp[1] = k * cart_topo.dims[1];
-        ompio_fh->f_comm->c_topo->topo.cart.cart_rank (ompio_fh->f_comm,coords_tmp,&tmp_rank);
-        ompio_fh->f_init_aggr_list[k] = tmp_rank; //change this to use get rank
-    }
-
-    //Initial Grouping
-    ompio_fh->f_init_procs_in_group = (int*)malloc (ompio_fh->f_init_procs_per_group * sizeof(int));
-    if (NULL == ompio_fh->f_init_procs_in_group) {
-        opal_output (1, "OUT OF MEMORY\n");
-        return OMPI_ERR_OUT_OF_RESOURCE;
-    }
-                                                                                                                                                              for (j=0 ; j< ompio_fh->f_size ; j++) {
-        ompio_fh->f_comm->c_topo->topo.cart.cart_coords (ompio_fh->f_comm, j, cart_topo.ndims, coords_tmp);
-	if (coords_tmp[0]  == cart_topo.coords[0]) {
-           if ((coords_tmp[1]/ompio_fh->f_init_procs_per_group) ==
-	       (cart_topo.coords[1]/ompio_fh->f_init_procs_per_group)) {
-
-	       ompio_fh->f_init_procs_in_group[n] = j;
-	       n++;
-	   }
-        }
-    }
-
-    /*print original group */
-    /*printf("RANK%d Initial distribution \n",ompio_fh->f_rank);
-    for(k = 0; k < ompio_fh->f_init_procs_per_group; k++){
-       printf("%d,", ompio_fh->f_init_procs_in_group[k]);
-    }
-    printf("\n");*/
-
-    if (NULL != cart_topo.dims) {
-       free (cart_topo.dims);
-       cart_topo.dims = NULL;
-    }
-    if (NULL != cart_topo.periods) {
-       free (cart_topo.periods);
-       cart_topo.periods = NULL;
-    }
-    if (NULL != cart_topo.coords) {
-       free (cart_topo.coords);
-       cart_topo.coords = NULL;
-    }
-
-    return OMPI_SUCCESS;
-}
